Struct ecs::scene::Scene

pub struct Scene<T: Entity> { /* fields omitted */ }

Scene is basically a manager for all entities and where to find them. It uses object pooling by instantiating a fixed number of entities at startup, in order to maintain a decent render speeds when creating and destoying entities during the game. Scene also provides tools for searching and borrowing spawned entities.

Implementations

impl<T: Entity> Scene<T>

pub fn new(size: usize, factories: Vec<Box<dyn Factory<T>>>) -> Self

Create a new Scene entity manager instance.

By setting the 'size' parameter, you preset the maximum amount of Object the new pool can hold and therefore spawn.

Factories are custom object factories that implement the Factory trait which can be called upon, by the Scene, when spawning a new object. Factories give the user the freedom to create different types of objects by customizing the output of these factories.

pub fn get_factory(&self, group: &Group) -> &Box<dyn Factory<T>>

pub fn mut_factory(&mut self, group: &Group) -> &mut Box<dyn Factory<T>>

pub fn list_spawned(&self) -> Vec<Spawn>

Returns a cloned list of spawn currently in use.

pub fn get_ref(&self, spawn: &Spawn) -> Ref<'_, T>

Returns a reference to a RefCell box containing the requested object. If the spawned object has been destroyed the inactive object will still be returned. You can use the methodes exists and exists_in_group to find out if objects are currently active.

pub fn get_mut(&self, spawn: &Spawn) -> RefMut<'_, T>

Same as the get_ref methode but returns a mutable reference.

pub fn test_all<P>(&self, predicate: &mut P) -> bool where
P: FnMut(&T) -> bool,

Run a custom test that tells if all active (spawned) objects comply to the predicate specified.

pub fn find_spawn(&self, name: &str) -> Option<Spawn>

Find an active (spawned) object by its spawn name.

pub fn find_spawn_in_group(&self, name: &str, group: Group) -> Option<Spawn>

Find an active (spawned) object by its spawn name and factory group. This methode can be faster as find_spawn, when there are multiple groups, sinds it does not need to itterate over all objects. Find_in_group can also come in handy when using the same name in different groups (sinds spawn names do not need to be unique).

pub fn search_components<P>(&self, predicate: P) -> Option<Spawn> where
P: FnMut(&T) -> bool,

As find_spawn, but lets you write a custom predicate using object values.

pub fn search_components_in_group<P>(
&self,
group: Group,
predicate: P
) -> Option<Spawn> where
P: FnMut(&T) -> bool,

As find_spawn_in_group, but lets you write a custom predicate using object values.

pub fn compare_against<F>(&self, against: Spawn, on_compare: F) -> Option<Spawn> where
F: FnMut(&T, &T) -> bool,

Compare an objects values, with the values of all other objects. Returns an Option of the Spawn on which the predicate succeeded first, or None is all comparisons failed.

pub fn compare_all<F>(&self, on_compare: F) -> Option<(Spawn, Spawn)> where
F: FnMut(&T, &T) -> bool,

Compares all values of all objects to eachother. Returns an Option of the two Spawn on which the predicate succeeded first, or None is all comparisons failed.

pub fn spawn(&mut self, name: &str, group: &Group) -> Result<Spawn, SceneError>

Spawn a new object. Spawned objects are updated every frame by the core ECS system. The spawn methode activates a new object that will inherit all the settings of the factory of the corresponding group. A name must be added to the spawn, this can be used to find the spawn if necessary.

pub fn destroy(&mut self, spawn: &Spawn)

Destroy an object. Destroy deactivates an object and therefore stops it from being updated by the core ECS system.

NOTE: Destroy is slow

pub fn wipe(&mut self, pointer: &Pointer)

pub fn exists(&self, spawn: &Spawn) -> bool

Checks if the object at the Pointer position has been spawned (is active).

pub fn exists_in_group(&self, spawn: &Spawn, group: Group) -> bool

Checks if the object with a specific group tag, and Pointer position has been spawned (is active).

Only check one group and, in case of many groups containing many objects, will therefore be faster than looping through all spawned objects.

pub fn size(&self) -> usize

Returns the maximum capacity of the pool.